+++ BioBloomCategorizer/MultiFilter.h
@@ -9,14 +9,14 @@
 #define MULTIFILTER_H_
 #include <string>
 #include <vector>
-#include "Common/BloomFilter.h"
+#include "btl_bloomfilter/BloomFilter.hpp"
 #include "boost/unordered/unordered_map.hpp"
 #include "boost/shared_ptr.hpp"
 using namespace std;
 
 class MultiFilter {
 public:
-	MultiFilter(uint16_t hashNum, uint16_t kmerSize);
+	MultiFilter(unsigned hashNum, unsigned kmerSize);
 	void addFilter(string const &filterID, boost::shared_ptr<BloomFilter> filter);
 	const boost::unordered_map<string, bool> multiContains(const unsigned char* kmer);
 	const boost::unordered_map<string, bool> multiContains(const unsigned char* kmer,
